feat: support attribute matching
Now giving a map in the query is interpreted as matching by the
attributes. All given attributes must match an attribute in the node.

This is simlar to CSS attribute selectors.

If the value in the map is a function, it is interpreted as a
predicate taht is called with the corresponding attribute
value. Otherwise an exact match is checked.